Add files via upload
Hey guys, I upload my files for part of the midweek causal inference question. 

The file "summary_stats.R" is the code I used to create the file "data_midweek_final.csv".

The file "midweek_analysis.R" is the code for the causal inference question. The only problem is that it seems that on the EPL, there is no difference (in descriptive statistics) in playing midweek or on the weekend. However, this varies over seasons. I also found that there may be some difference for playing in a "short week" (i.e. having two or more games in the same week). 

Let me know what you think, specially with the first R file, since there may be some mistakes on it.